APIルートの認証ミドルウェアを追加し、ルート設定を整理。APIクライアントの設定を更新し、認証機能を強化。
