+++ apps/server/src/app.ts
@@ -2,4 +2,4 @@ export { hc } from "hono/client";
 
 export { handle } from "hono/vercel";
 
-export { app } from "./routes";
+export { apiRoutes } from "./routes";
