+++ apps/server/src/routes/routes.ts
@@ -1,0 +1,14 @@
+import { authHandler, verifyAuth } from "@hono/auth-js";
+import { Hono } from "hono";
+import { authMiddleware } from "../middlewares/auth";
+import { exampleRoutes } from "./example";
+
+const app = new Hono();
+
+app.use("*", authMiddleware);
+
+app.use("/api/auth/*", authHandler());
+
+app.use("/api/*", verifyAuth());
+
+export const apiRoutes = app.basePath("/api").route("/examples", exampleRoutes);
